MEPA: Remove VTSS_FEATURE_EDC_FW_LOAD option and code
The VTSS_FEATURE_EDC_FW_LOAD was used in header files being part of
MEPA, but MEPA is not setting these defines as it is targeting an ABI
stable interfaces. This is therefore causing ABI issues.

Earier varsion of #267 suggest to disable it. Here we remove the feature
as it cannot be enabled.

Review-record: https://bitbucket.microchip.com/projects/UNGE/repos/sw-mesa/pull-requests/281